### 2025-01-15 e0cbffe
Adding missing acceleration file and updated plot settings to make them look a bit nicer
