### 2025-01-16 4ae5de8
JOSS paper comments
Added DOIs to reference list, spelled out USA
