+++ paper.bib
@@ -4,6 +4,7 @@
   volume={19},
   year={2010},
   publisher={{SIAM}}
+  doi = {10.1137/1.9780898718577}
 }
 
 @article{zarrugh:1974,
@@ -15,6 +16,7 @@
   pages={293--306},
   year={1974},
   publisher={Springer}
+  doi = {10.1007/BF00430237}
 }
 
 @article{seth:2011,
@@ -25,6 +27,7 @@
   pages={212--232},
   year={2011},
   publisher={Elsevier}
+  doi = {10.1016/j.piutam.2011.04.021}
 }
 
 @article{ackermann:2010,
@@ -36,6 +39,7 @@
   pages={1055--1060},
   year={2010},
   publisher={Elsevier}
+  doi = {10.1016/j.jbiomech.2009.12.012}
 }
 
 @article{nitschke:2023,
@@ -46,6 +50,7 @@
   pages={e14852},
   year={2023},
   publisher={PeerJ Inc.}
+  doi = {10.7717/peerj.14852}
 }
 
 @article{koelewijn:2018,
@@ -57,6 +62,7 @@
   pages={521--531},
   year={2018},
   publisher={Taylor \& Francis}
+  doi = {10.1080/10255842.2018.1490954}
 }
 
 @article{nitschke:2020,
@@ -68,6 +74,7 @@
   pages={17655},
   year={2020},
   publisher={Nature Publishing Group UK London}
+  doi = {10.1038/s41598-020-73856-w}
 }
 
 @article{nitschke:2024,
@@ -78,6 +85,7 @@
   pages={1285845},
   year={2024},
   publisher={Frontiers Media SA}
+  doi = {10.3389/fbioe.2024.1285845}
 }
 
 @article{dorscky:2019a,
@@ -98,6 +106,7 @@
   pages={1931--1940},
   year={2022},
   publisher={IEEE}
+  doi = {10.1109/TNSRE.2022.3189038}
 }
 
 @article{Geyer:2010,
@@ -109,6 +118,7 @@
   pages={263--273},
   year={2010},
   publisher={IEEE}
+  doi = {10.1109/tnsre.2010.2047592}
 }
 
 @article{Wachter:2006,
@@ -120,6 +130,7 @@
   pages={25--57},
   year={2006},
   publisher={Springer}
+  doi = {10.1007/s10107-004-0559-y }
 }
 
 @article{koelewijn:2016,
@@ -130,6 +141,7 @@
   pages={219--225},
   year={2016},
   publisher={Elsevier}
+  doi = {10.1016/j.gaitpost.2016.07.007}
 }
 
 @article{dorschkynitschke:2024,
@@ -139,6 +151,7 @@
   pages={2023--05},
   year={2024},
   publisher={Cold Spring Harbor Laboratory}
+  doi = {10.1101/2023.05.25.542228}
 }
 
 @article{dorschky:2019b,
@@ -150,6 +163,7 @@
   pages={869--879},
   year={2019},
   publisher={Taylor \& Francis}
+  doi = {10.1080/10255842.2019.1601179}
 }
 
 @article{falisse:2019,
@@ -161,4 +175,5 @@
   pages={20190402},
   year={2019},
   publisher={The Royal Society}
+  doi = {10.1098/rsif.2019.0402}
 }
