+++ docs/Coding_Guidelines.md
@@ -1,0 +1,446 @@
+# BioMAC-Sim-Toolbox: Coding Guidelines
+
+This coding guidelines explain applied naming conventions and implementation principles. At the moment not all guidelines were applied in the code. Feel free to change and improve the code ;)
+
+## Folders and Files
+The folder structure of the protect is currently the following:
+
+````
+|-- project home directory
+|   |-- docs
+|   |   |-- Doxygen
+|   |   |-- files
+|   |-- src
+|   |   |-- HelperFunctions
+|   |   |-- model
+|   |   |   |-- gait2dc
+|   |   |   |   |-- ...
+|   |   |   |-- gait2d_osim
+|   |   |   |   |-- ...
+|   |   |   |-- gait3d
+|   |   |   |   |-- ...
+|   |   |   |-- Model.m
+|   |   |-- problem
+|   |   |   |-- @Collocation
+|   |   |   |-- Problem.m
+|   |   |   |-- ...
+|   |   |-- result
+|   |   |   |-- ...
+|   |   |-- scripts
+|   |   |   |-- ...
+|   |   |-- solver
+|   |   |   |-- IPOPT.m
+|   |   |   |-- Solver.m
+|   |   |-- tests
+|   |   |   |-- Gait2dcTest.m
+|   |   |   |-- gait3dTest.m
+|   |   |   |-- ModelTest.m
+|   |   |   |-- ...
+|   |   |-- Toolboxes
+|   |   |-- trackingData
+|   |   |   |-- @TrackingData
+|   |   |-- ...
+|	|-- ...
+````
+
+
+In Matlab, classes can have a own class folder named `@<ClassName>`. This folder contains the file defining the model class (e.g.: Gait2dc.m) and other functions of this class (e.g.: simuAccGyro.m). This other functions have to be declared in the class file but the content is implemented in a separate file.
+
+
+## Classes
+- Starting with uppercase
+- E.g.: `Model`, `Problem`, `Collocation`, ...
+
+## Functions
+- Starting with lowercase
+- New word starts with uppercase
+- `initModel()`, `initMex()`, `solve()`, ...
+- Exception: Constructors
+
+## Properties
+- Starting with lowercase
+- Numbers:
+	- `n<Name>`
+	- E.g.: Number of `dofs` is `nDofs`
+- Single index:
+	- Used in loops
+	- `i<Name>`
+	- E.g.: 
+		```matlab
+		for iDof = 1 : obj.nDofs
+			disp(num2str(iDof));
+		end
+		```
+- Multiple indices in an array:
+	- Parameter
+	- `idx<Name>`
+	- E.g.: `idxArmdof`
+- Hidden properties:
+	- Used to save the content of dependent variables
+	- `h<Name>`
+	- E.g.: `hidxArmdof` for `idxArmdof`
+- Constants:
+	- Uppercase
+	- E.g.: `GRAVITY` in the Gait3d-Class
+
+## Default Values
+- Should be set in property definition:
+	```matlab
+	properties (SetObservable, AbortSet, SetAccess = protected)
+		%> Bodyheight in m (default: 1.80)
+		bodyheight = 1.80
+	end
+	```
+- Or it should be set in the constructor:
+	```matlab
+	function [obj] = IPOPT()
+		% Set default options for termination
+		obj.solverOptions.tol = 0.0002; % Desired convergence tolerance
+		obj.solverOptions.max_iter = 4000; % Maximum number of iterations 
+		% ...
+	end
+	```
+- Do not change them in the code!
+- Use setter function to change the value after object creation. If needed, write an own setter function.
+
+## Inheritance
+- Use inheritance instead of changing the functionality of the basic class.
+- E.g.: 
+	````
+	handle
+	|--Model
[…]
