### 2025-01-25 bfcf7a9
Reorganized the build process for gait2dc model
The original version relied on acc_al.c which could not be built from acc.al.  Also renamed some of the source files for consistency.
After these changes, IMUTracking2D.scriptIMU2D still works correctly, so everything is probably good.
