+++ src/model/gait2dc/c_files/gait2dc_FK_al.c
@@ -1,0 +1,1067 @@
+// This file was generated by autolevclean.c and contains C code generated by Autolev
+
+#include <math.h>
+#include "gait2dc.h"
+void gait2dc_FK_al(param_struct* par, double q[NDOF], double qd[NDOF], double qdd[NDOF],
+   double fk[NFK], double dfk_dq[NFK][NDOF],
+   double fkdot[NFK], double dfkdot_dq[NFK][NDOF]) {
+	double q1 = q[0];
+	double q1p = qd[0];
+	double q1pp = qdd[0];
+	double q2 = q[1];
+	double q2p = qd[1];
+	double q2pp = qdd[1];
+	double q3 = q[2];
+	double q3p = qd[2];
+	double q3pp = qdd[2];
+	double q4 = q[3];
+	double q4p = qd[3];
+	double q4pp = qdd[3];
+	double q5 = q[4];
+	double q5p = qd[4];
+	double q5pp = qdd[4];
+	double q6 = q[5];
+	double q6p = qd[5];
+	double q6pp = qdd[5];
+	double q7 = q[6];
+	double q7p = qd[6];
+	double q7pp = qdd[6];
+	double q8 = q[7];
+	double q8p = qd[7];
+	double q8pp = qdd[7];
+	double q9 = q[8];
+	double q9p = qd[8];
+	double q9pp = qdd[8];
+	static double z[199];
+  z[1] = cos(q3);
+  z[2] = sin(q3);
+  z[3] = cos(q4);
+  z[4] = sin(q4);
+  z[5] = cos(q5);
+  z[6] = sin(q5);
+  z[7] = z[1]*z[3] - z[2]*z[4];
+  z[8] = -z[1]*z[4] - z[2]*z[3];
+  z[9] = z[1]*z[4] + z[2]*z[3];
+  z[12] = cos(q6);
+  z[13] = sin(q6);
+  z[14] = z[5]*z[9] + z[6]*z[7];
+  z[15] = z[5]*z[7] + z[6]*z[8];
+  z[17] = z[5]*z[7] - z[6]*z[9];
+  z[18] = z[5]*z[8] - z[6]*z[7];
+  z[25] = cos(q7);
+  z[26] = sin(q7);
+  z[27] = cos(q8);
+  z[28] = sin(q8);
+  z[29] = z[1]*z[25] - z[2]*z[26];
+  z[30] = -z[1]*z[26] - z[2]*z[25];
+  z[31] = z[1]*z[26] + z[2]*z[25];
+  z[34] = cos(q9);
+  z[35] = sin(q9);
+  z[36] = z[27]*z[31] + z[28]*z[29];
+  z[37] = z[27]*z[29] + z[28]*z[30];
+  z[39] = z[27]*z[29] - z[28]*z[31];
+  z[40] = z[27]*z[30] - z[28]*z[29];
+  z[47] = z[12]*z[15] + z[13]*z[18];
+  z[48] = z[12]*z[18] - z[13]*z[15];
+  z[49] = z[12]*z[14] + z[13]*z[17];
+  z[50] = z[12]*z[17] - z[13]*z[14];
+  z[51] = z[34]*z[37] + z[35]*z[40];
[…]
