+++ src/model/gait2dc/@Gait2dc/getMexFiles.m
@@ -5,7 +5,7 @@
 %> Details: Gait2dc::getMexFiles()
 %>
 %> @author Ton van den Bogert
-%> @date January, 2018
+%> @date January, 2025
 %======================================================================
 
 %======================================================================
@@ -14,149 +14,155 @@
 %> @public
 %>
 %> @details
-%> Generates Autolev source file for model and builds the MEX function
-%>
-%> @param  mexUpdate  (optional) Bool: If true, MEX will be updated using
-%>                    the Makefile (default: 0)
+%> Builds the MEX function from Autolev source code, and from gait2dc.c MEX wrapper.
 %======================================================================
-function getMexFiles(mexUpdate)
+function getMexFiles()
 
-clear mex						% to make sure mex files are not loaded, so we can delete/overwrite
-
-% change directory
-currentFolder = pwd;
-methodFolder = strrep(mfilename('fullpath'), [filesep , mfilename], '');
-cfilesFolder = strrep(methodFolder, '@Gait2dc', 'c_files');
-cd(cfilesFolder);
-
-if isdeployed % don't compile if function is running in deployed mode (e.g. on HPC cluster)
-    return
-end
-
-if nargin > 0 && mexUpdate % update the mex file
+    clear mex % to make sure mex files are not loaded, so we can delete/overwrite if needed
     
-    if ismac
-        if (system('make -f Makefile.mac') ~= 0)
-            error('Error in compilation of MEX function.');
-        end
-    elseif ispc
-        if (system('@echo off & call "vcvars32" & nmake /nologo /F Makefile.win') ~= 0)
-            %cd ..
-            error('Error in compilation of MEX function.');
-        end
-    else
-        error('Ubuntu is currently not supported');
+    % change directory
+    currentFolder = pwd;
+    methodFolder = strrep(mfilename('fullpath'), [filesep , mfilename], '');
+    cfilesFolder = strrep(methodFolder, '@Gait2dc', 'c_files');
+    cd(cfilesFolder);
+    
+    if isdeployed % don't compile if function is running in deployed mode (e.g. on HPC cluster)
+        return
     end
     
-else  % create new mex
-    mexfile = ['gait2dc.' mexext];
-    % Try to run Autolev if the raw C files are outdated
-    if needs_rebuild('contact_al_raw.c',{'contact.al'})
+    % Run Autolev (if installed) to rebuild any C files that are outdated
+    % (Note, we don't need to run gait2dc_FK.al here, it will be done within
+    % gait2dc_dyn.al)
+    clean_needed = false;
+    if needs_rebuild('contact_al.c',{'contact.al', 'extforce.a'})
         runautolev('contact');
+        clean_needed = true;
     end
-    if needs_rebuild('gait2dc_dynamics_al_raw.c',{'gait2dc_dynamics.al', 'gait2dc_FK.al'})
-        runautolev('gait2dc_dynamics');
+
+    if needs_rebuild('gait2dc_dyn_al.c',{'gait2dc_dyn.al', 'gait2dc_FK.al', 'gait2dc_kin.al'})
+        runautolev('gait2dc_dyn');
+        clean_needed = true;
     end
-    if needs_rebuild('gait2dc_stick_al_raw.c',{'gait2dc_stick.al', 'gait2dc_FK.al'})
+    if needs_rebuild('gait2dc_acc_al.c',{'gait2dc_acc.al', 'accelerometer.a', 'gait2dc_kin.al'})
+        runautolev('gait2dc_acc');
+        clean_needed = true;
+    end
+    if needs_rebuild('gait2dc_stick_al.c',{'gait2dc_stick.al', 'gait2dc_kin.al'})
         runautolev('gait2dc_stick');
+        clean_needed = true;
     end
[…]
