+++ src/model/gait2dc/c_files/gait2dc_FK_al_raw.c
@@ -1,1169 +1,0 @@
-/* The name of this program is gait2dc_FK_al_raw.c */
-/* Created by Autolev 4.1 on Thu Oct 24 13:28:07 2024 */
-
-#include <ctype.h> 
-#include <math.h>  
-#include <stdarg.h>
-#include <stdio.h> 
-#include <stdlib.h>
-#include <string.h>
-#define  _NAN      9.99999999999999E+305
-
-void     eqns1     (void);
-void     output    (FILE *Fptr[] );
-void     readf     (FILE *Fp, double *next, ...);
-void     pgets     (FILE *Fp, double *x);
-void     writef    (FILE *Fp, char format[], ...);
-
-double   par__ShankLen,par__ThighLen,q1,q2,q3,q4,q5,q6,q7,q8,q9,q1p,q2p,q3p,q4p,
-  q5p,q6p,q7p,q8p,q9p;
-double   Pi,DEGtoRAD,RADtoDEG,z[245],fk[42],dfk_dq[42][9],fkdot[42],dfkdot_dq[
-  42][9],Encode[4];
-
-/* ................................ MAIN ............................. */
-int      main         (void)
-{
-FILE     *Fptr[1];
-int      iloop;
-
-/* Open input and output files */
-for(iloop=0;  iloop<=0;  iloop++)
-  {
-  char fileName[256];
-  if( !iloop ) strcpy(fileName, "gait2dc_FK_al_raw.in");
-  else sprintf(fileName, "gait2dc_FK_al_raw.%d", iloop);
-  if( (Fptr[iloop] = fopen(fileName, iloop ? "w" : "r")) == NULL)
-    {printf("Error: unable to open file %s\n", fileName);  exit(0);}
-  }
- 
-/* Read top of input file */
-for(iloop=0;  iloop<6;  iloop++) pgets(Fptr[0],NULL);
-
-/* Read values of constants from input file */
-readf(Fptr[0],&par__ShankLen,&par__ThighLen,&q1,&q2,&q3,&q4,&q5,&q6,&q7,&q8,&q9,&
-  q1p,&q2p,&q3p,&q4p,&q5p,&q6p,&q7p,&q8p,&q9p,NULL);
-
-/* Write heading(s) to output file(s) */
-
-/* Unit conversions */
-  Pi       = 3.141592653589793;
-  DEGtoRAD = Pi/180.0;
-  RADtoDEG = 180.0/Pi;
-
-/* Evaluate output quantities */
-eqns1();
-output( Fptr );
-
-/* Inform user of input and output filename(s) */
-puts( "\n Input is in the file gait2dc_FK_al_raw.in" );
-return 0;
-}
-
-
-/* ................................ EQNS1 ............................. */
-void     eqns1        (void)
-{
-
-/* Evaluate constants */
-  z[1] = cos(q3);
-  z[2] = sin(q3);
[…]

[…]
