### 2025-01-27 076ed28
tvdbogert-issue-#6 (#9)
* deleted gait2dc/c_files/Makefile.win

The build logic is all coded in gait2dc/@Gait2dc/getMexFiles.m, so we don't need a Makefile.

* Reorganized the build process for gait2dc model

The original version relied on acc_al.c which could not be built from acc.al.  Also renamed some of the source files for consistency.
After these changes, IMUTracking2D.scriptIMU2D still works correctly, so everything is probably good.
