### 2025-01-27 1e71388
Removing unnecessary code
