+++ src/model/gait2dc/c_files/gait2dc_dynamics_al.c
@@ -1,0 +1,2334 @@
+// This file was generated by autolevclean.c and contains C code generated by Autolev
+
+#include <math.h>
+#include "gait2dc.h"
+void gait2dc_dynamics_al(param_struct* par, double q[NDOF], double qd[NDOF], double qdd[NDOF],
+   double mom[NMOM], double grf[NGRF], double Zero[NDOF], double dz_dq[NDOF][NDOF],
+   double dz_dqd[NDOF][NDOF], double dz_dqdd[NDOF][NDOF],
+   double dz_dmom[NDOF][NMOM], double dz_dgrf[NDOF][NGRF],
+   double fk[NFK], double dfk_dq[NFK][NDOF],
+   double fkdot[NFK], double dfkdot_dq[NFK][NDOF]) {
+	double q1 = q[0];
+	double q1p = qd[0];
+	double q1pp = qdd[0];
+	double q2 = q[1];
+	double q2p = qd[1];
+	double q2pp = qdd[1];
+	double q3 = q[2];
+	double q3p = qd[2];
+	double q3pp = qdd[2];
+	double q4 = q[3];
+	double q4p = qd[3];
+	double q4pp = qdd[3];
+	double q5 = q[4];
+	double q5p = qd[4];
+	double q5pp = qdd[4];
+	double q6 = q[5];
+	double q6p = qd[5];
+	double q6pp = qdd[5];
+	double q7 = q[6];
+	double q7p = qd[6];
+	double q7pp = qdd[6];
+	double q8 = q[7];
+	double q8p = qd[7];
+	double q8pp = qdd[7];
+	double q9 = q[8];
+	double q9p = qd[8];
+	double q9pp = qdd[8];
+	double mom1 = mom[0];
+	double mom2 = mom[1];
+	double mom3 = mom[2];
+	double mom4 = mom[3];
+	double mom5 = mom[4];
+	double mom6 = mom[5];
+	double RFootFx = grf[0];
+	double RFootFy = grf[1];
+	double RFootMz = grf[2];
+	double LFootFx = grf[3];
+	double LFootFy = grf[4];
+	double LFootMz = grf[5];
+	static double z[634];
+  z[1] = cos(q3);
+  z[2] = sin(q3);
+  z[3] = par->TrunkCMy*q3p;
+  z[4] = q3p*z[3];
+  z[5] = cos(q4);
+  z[6] = sin(q4);
+  z[7] = z[1]*z[5] - z[2]*z[6];
+  z[8] = -z[1]*z[6] - z[2]*z[5];
+  z[9] = z[1]*z[6] + z[2]*z[5];
+  z[10] = par->ThighCMy*(q3p+q4p);
+  z[11] = q3p + q4p;
+  z[12] = z[10]*z[11];
+  z[13] = cos(q5);
+  z[14] = sin(q5);
+  z[15] = par->ThighLen*(q3p+q4p);
+  z[16] = z[11]*z[15];
+  z[17] = z[7]*z[13] + z[8]*z[14];
+  z[18] = z[8]*z[13] - z[7]*z[14];
+  z[19] = z[7]*z[14] + z[9]*z[13];
+  z[20] = z[7]*z[13] - z[9]*z[14];
+  z[21] = par->ThighLen*z[13];
+  z[22] = par->ThighLen*z[14];
+  z[23] = z[21] - par->ShankCMy;
+  z[24] = z[14]*z[16];
[…]
