+++ paper.bib
@@ -3,7 +3,7 @@
   author={Betts, JT},
   volume={19},
   year={2010},
-  publisher={{SIAM}}
+  publisher={{SIAM}},
   doi = {10.1137/1.9780898718577}
 }
 
@@ -15,7 +15,7 @@
   number={4},
   pages={293--306},
   year={1974},
-  publisher={Springer}
+  publisher={Springer},
   doi = {10.1007/BF00430237}
 }
 
@@ -26,7 +26,7 @@
   volume={2},
   pages={212--232},
   year={2011},
-  publisher={Elsevier}
+  publisher={Elsevier},
   doi = {10.1016/j.piutam.2011.04.021}
 }
 
@@ -38,7 +38,7 @@
   number={6},
   pages={1055--1060},
   year={2010},
-  publisher={Elsevier}
+  publisher={Elsevier},
   doi = {10.1016/j.jbiomech.2009.12.012}
 }
 
@@ -49,7 +49,7 @@
   volume={11},
   pages={e14852},
   year={2023},
-  publisher={PeerJ Inc.}
+  publisher={PeerJ Inc.},
   doi = {10.7717/peerj.14852}
 }
 
@@ -61,7 +61,7 @@
   number={8},
   pages={521--531},
   year={2018},
-  publisher={Taylor \& Francis}
+  publisher={Taylor \& Francis},
   doi = {10.1080/10255842.2018.1490954}
 }
 
@@ -73,7 +73,7 @@
   number={1},
   pages={17655},
   year={2020},
-  publisher={Nature Publishing Group UK London}
+  publisher={Nature Publishing Group UK London},
   doi = {10.1038/s41598-020-73856-w}
 }
 
@@ -84,7 +84,7 @@
   volume={12},
   pages={1285845},
   year={2024},
-  publisher={Frontiers Media SA}
+  publisher={Frontiers Media SA},
   doi = {10.3389/fbioe.2024.1285845}
 }
 
@@ -105,7 +105,7 @@
   volume={30},
   pages={1931--1940},
   year={2022},
-  publisher={IEEE}
+  publisher={IEEE},
   doi = {10.1109/TNSRE.2022.3189038}
 }
 
@@ -117,7 +117,7 @@
   number={3},
   pages={263--273},
   year={2010},
-  publisher={IEEE}
+  publisher={IEEE},
   doi = {10.1109/tnsre.2010.2047592}
 }
 
@@ -129,7 +129,7 @@
   number={1},
   pages={25--57},
   year={2006},
-  publisher={Springer}
+  publisher={Springer},
   doi = {10.1007/s10107-004-0559-y }
 }
 
@@ -140,7 +140,7 @@
   volume={49},
   pages={219--225},
   year={2016},
-  publisher={Elsevier}
+  publisher={Elsevier},
   doi = {10.1016/j.gaitpost.2016.07.007}
 }
 
@@ -150,7 +150,7 @@
   journal={bioRxiv},
   pages={2023--05},
   year={2024},
-  publisher={Cold Spring Harbor Laboratory}
+  publisher={Cold Spring Harbor Laboratory},
   doi = {10.1101/2023.05.25.542228}
 }
 
@@ -162,7 +162,7 @@
   number={8},
   pages={869--879},
   year={2019},
-  publisher={Taylor \& Francis}
+  publisher={Taylor \& Francis},
   doi = {10.1080/10255842.2019.1601179}
 }
 
@@ -174,6 +174,6 @@
   number={157},
   pages={20190402},
   year={2019},
-  publisher={The Royal Society}
+  publisher={The Royal Society},
   doi = {10.1098/rsif.2019.0402}
 }
