### 2025-01-27 ac7054b
Fixing issues (#10)
* Fixing inconsistencies found during Innsbruck visit

* JOSS paper comments

Added DOIs to reference list, spelled out USA

* Removing unnecessary code

* Added commas before DOIs
