+++ paper.bib
@@ -3,7 +3,8 @@
   author={Betts, JT},
   volume={19},
   year={2010},
-  publisher={{SIAM}}
+  publisher={{SIAM}},
+  doi = {10.1137/1.9780898718577}
 }
 
 @article{zarrugh:1974,
@@ -14,7 +15,8 @@
   number={4},
   pages={293--306},
   year={1974},
-  publisher={Springer}
+  publisher={Springer},
+  doi = {10.1007/BF00430237}
 }
 
 @article{seth:2011,
@@ -24,7 +26,8 @@
   volume={2},
   pages={212--232},
   year={2011},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi = {10.1016/j.piutam.2011.04.021}
 }
 
 @article{ackermann:2010,
@@ -35,7 +38,8 @@
   number={6},
   pages={1055--1060},
   year={2010},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi = {10.1016/j.jbiomech.2009.12.012}
 }
 
 @article{nitschke:2023,
@@ -45,7 +49,8 @@
   volume={11},
   pages={e14852},
   year={2023},
-  publisher={PeerJ Inc.}
+  publisher={PeerJ Inc.},
+  doi = {10.7717/peerj.14852}
 }
 
 @article{koelewijn:2018,
@@ -56,7 +61,8 @@
   number={8},
   pages={521--531},
   year={2018},
-  publisher={Taylor \& Francis}
+  publisher={Taylor \& Francis},
+  doi = {10.1080/10255842.2018.1490954}
 }
 
 @article{nitschke:2020,
@@ -67,7 +73,8 @@
   number={1},
   pages={17655},
   year={2020},
-  publisher={Nature Publishing Group UK London}
+  publisher={Nature Publishing Group UK London},
+  doi = {10.1038/s41598-020-73856-w}
 }
 
 @article{nitschke:2024,
@@ -77,7 +84,8 @@
   volume={12},
   pages={1285845},
   year={2024},
-  publisher={Frontiers Media SA}
+  publisher={Frontiers Media SA},
+  doi = {10.3389/fbioe.2024.1285845}
 }
 
 @article{dorscky:2019a,
@@ -97,7 +105,8 @@
   volume={30},
   pages={1931--1940},
   year={2022},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi = {10.1109/TNSRE.2022.3189038}
 }
 
 @article{Geyer:2010,
@@ -108,7 +117,8 @@
   number={3},
   pages={263--273},
   year={2010},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi = {10.1109/tnsre.2010.2047592}
 }
 
 @article{Wachter:2006,
@@ -119,7 +129,8 @@
   number={1},
   pages={25--57},
   year={2006},
-  publisher={Springer}
+  publisher={Springer},
+  doi = {10.1007/s10107-004-0559-y }
 }
 
 @article{koelewijn:2016,
@@ -129,7 +140,8 @@
   volume={49},
   pages={219--225},
   year={2016},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi = {10.1016/j.gaitpost.2016.07.007}
 }
 
 @article{dorschkynitschke:2024,
@@ -138,7 +150,8 @@
   journal={bioRxiv},
   pages={2023--05},
   year={2024},
-  publisher={Cold Spring Harbor Laboratory}
+  publisher={Cold Spring Harbor Laboratory},
+  doi = {10.1101/2023.05.25.542228}
 }
 
 @article{dorschky:2019b,
@@ -149,7 +162,8 @@
   number={8},
[…]
