+++ src/tests/Gait2dcTest.m
@@ -18,27 +18,26 @@
 %> @endcode
 %>
 %> Overview:
-%> | Testname/Tag                    | Apparel | Derivative | Freefall | Isometric | Memory | Speed | Neutral | getDynamics | getGRF | getJointmoments | simuAccGyro | showStick |
-%> |---------------------------------|---------|------------|----------|-----------|--------|-------|---------|-------------|--------|-----------------|-------------|-----------|
-%> | derivativetest_Dynamics         |         | x          |          |           |        |       |         | x           |        |                 |             |           |
-%> | derivativetest_SimuAccGyro_Acc  |         | x          |          |           |        |       |         |             |        |                 | x           |           |
-%> | derivativetest_SimuAccGyro_Gyro |         | x          |          |           |        |       |         |             |        |                 | x           |           |
-%> | derivativetest_Moments          |         | x          |          |           |        |       |         |             |        | x               |             |           |
-%> | derivativetest_GRF              |         | x          |          |           |        |       |         |             | x      |                 |             |           |
-%> | test_simulateFreefall           |         |            | x        |           |        |       |         | x           |        |                 |             |           |
-%> | test_showStickNeutral           |         |            |          |           |        |       | x       |             |        |                 |             | x         |
-%> | test_dynamics                   |         |            |          |           |        |       |         | x           | x      | x               |             |           |
-%> | test_memory                     |         |            |          |           | x      |       |         | x           |        |                 |             |           |
-%> | test_speedOfSimuAccGyro         |         |            |          |           |        | x     |         |             |        |                 | x           |           |
-%> | test_speedOfMex                 |         |            |          |           |        | x     |         | x           |        |                 |             |           |
-%> | test_s_hip_flexion_gyro         |         |            |          |           |        |       |         |             |        |                 | x           |           |
-%> | test_s_static_upright_acc       |         |            |          |           |        |       | x       |             |        |                 | x           |           |
-%> | test_s_static_upright_gyro      |         |            |          |           |        |       | x       |             |        |                 | x           |           |
-%> | test_dynamicGRF                 |         |            |          |           |        |       |         |             | x      |                 |             |           |
-%> | test_isokineticMuscles          |         |            |          |           |        |       | x       | x           |        | x               |             |           |
-%> | test_isometricMuscles           |         |            |          | x         |        |       |         | x           |        | x               |             |           |
-%> | test_apparel                    | x       |            |          |           |        |       |         | x           |        | x               |             |           |
-%> | test_grf                        |         |            |          |           |        |       |         |             | x      |                 |             |           |
+%> | Testname/Tag                    | Derivative | Freefall | Isometric | Memory | Speed | Neutral | getDynamics | getGRF | getJointmoments | simuAccGyro | showStick |
+%> |---------------------------------|------------|----------|-----------|--------|-------|---------|-------------|--------|-----------------|-------------|-----------|
+%> | derivativetest_Dynamics         | x          |          |           |        |       |         | x           |        |                 |             |           |
+%> | derivativetest_SimuAccGyro_Acc  | x          |          |           |        |       |         |             |        |                 | x           |           |
+%> | derivativetest_SimuAccGyro_Gyro | x          |          |           |        |       |         |             |        |                 | x           |           |
+%> | derivativetest_Moments          | x          |          |           |        |       |         |             |        | x               |             |           |
[…]
