### 2025-01-27 293117f
Deleted gait2dc_dynamics_al.c
This file is now named gait2dc_dyn_al.c, which is already in the repository.
getMexFiles.m no longer uses gait2dc_dynamics_al.c.

Fixes #11
