### 2025-01-28 d2e2b0e
Update build_docs.sh
Fixing doxygen?
