### 2025-02-06 #7642 @Kevin-Mattheus-Moerman
@adkoele okay thanks for the update. If you can make these changes in a timely manor we can wait. However if you think you need more than a month then I suggest we retract this submission at this point and that you resubmit when the changes are implemented. Let me know what you think.

Fyi my GIBBON project features basic automated testing (a script triggering the test scripts). If you need an example.
