+++ .github/workflows/main.yml
@@ -23,12 +23,11 @@ jobs:
     - uses: actions/checkout@v2
         
     - name: Doxygen Action
-      uses: mattnotmitt/doxygen-action@v1.1.0
+      uses: DenverCoder1/doxygen-github-pages-action@v2.0.0
       with:
-        # Path to Doxyfile
-        doxyfile-path: "./Doxyfile" # default is ./Doxyfile
-        # Working directory
-        working-directory: "./docs/Doxygen" # default is .
+          github_token: ${{ secrets.GITHUB_TOKEN }}
+          folder: ./docs/Doxygen/html
+          config_file: ./docs/Doxygen/Doxyfile
     
     - name: Upload static files as artifact
       id: deployment
