+++ .github/workflows/main.yml
@@ -35,7 +35,7 @@ jobs:
       id: deployment
       uses: actions/upload-pages-artifact@v3 # or specific "vX.X.X" version tag for this action
       with:
-        path: ./docs/Doxygen/Doxygen_HTML
+        path: ./docs/Doxygen/html
 
     # Deployment job
   deploy:
