+++ paper.bib
@@ -1,0 +1,179 @@
+@book{betts:2010,
+  title={Practical methods for optimal control and estimation using nonlinear programming},
+  author={Betts, JT},
+  volume={19},
+  year={2010},
+  publisher={{SIAM}},
+  doi = {10.1137/1.9780898718577}
+}
+
+@article{zarrugh:1974,
+  title={Optimization of energy expenditure during level walking},
+  author={Zarrugh, MY and Todd, FN and Ralston, HJ},
+  Journal={European Journal of Applied Physiology and Occupational Physiology},
+  volume={33},
+  number={4},
+  pages={293--306},
+  year={1974},
+  publisher={Springer},
+  doi = {10.1007/BF00430237}
+}
+
+@article{seth:2011,
+  title={OpenSim: a musculoskeletal modeling and simulation framework for in silico investigations and exchange},
+  author={Seth, Ajay and Sherman, Michael and Reinbolt, Jeffrey A and Delp, Scott L},
+  journal={Procedia Iutam},
+  volume={2},
+  pages={212--232},
+  year={2011},
+  publisher={Elsevier},
+  doi = {10.1016/j.piutam.2011.04.021}
+}
+
+@article{ackermann:2010,
+  title={Optimality principles for model-based prediction of human gait},
+  author={Ackermann, Marko and Van den Bogert, Antonie J},
+  journal={Journal of biomechanics},
+  volume={43},
+  number={6},
+  pages={1055--1060},
+  year={2010},
+  publisher={Elsevier},
+  doi = {10.1016/j.jbiomech.2009.12.012}
+}
+
+@article{nitschke:2023,
+  title={Change the direction: 3D optimal control simulation by directly tracking marker and ground reaction force data},
+  author={Nitschke, Marlies and Marzilger, Robert and Leyendecker, Sigrid and Eskofier, Bjoern M and Koelewijn, Anne D},
+  journal={PeerJ},
+  volume={11},
+  pages={e14852},
+  year={2023},
+  publisher={PeerJ Inc.},
+  doi = {10.7717/peerj.14852}
+}
+
+@article{koelewijn:2018,
+  title={A metabolic energy expenditure model with a continuous first derivative and its application to predictive simulations of gait},
+  author={Koelewijn, Anne D and Dorschky, Eva and Van den Bogert, Antonie J},
+  journal={Computer methods in biomechanics and biomedical engineering},
+  volume={21},
+  number={8},
+  pages={521--531},
+  year={2018},
+  publisher={Taylor \& Francis},
+  doi = {10.1080/10255842.2018.1490954}
+}
+
+@article{nitschke:2020,
+  title={Efficient trajectory optimization for curved running using a 3D musculoskeletal model with implicit dynamics},
+  author={Nitschke, Marlies and Dorschky, Eva and Heinrich, Dieter and Schlarb, Heiko and Eskofier, Bjoern M and Koelewijn, Anne D and van den Bogert, Antonie J},
+  journal={Scientific reports},
+  volume={10},
+  number={1},
+  pages={17655},
+  year={2020},
+  publisher={Nature Publishing Group UK London},
+  doi = {10.1038/s41598-020-73856-w}
+}
+
+@article{nitschke:2024,
+  title={Estimating 3D kinematics and kinetics from virtual inertial sensor data through musculoskeletal movement simulations},
+  author={Nitschke, Marlies and Dorschky, Eva and Leyendecker, Sigrid and Eskofier, Bjoern M and Koelewijn, Anne D},
+  journal={Frontiers in Bioengineering and Biotechnology},
+  volume={12},
+  pages={1285845},
+  year={2024},
[…]
