+++ paper.md
@@ -1,0 +1,134 @@
+---
+title: 'BioMAC-Sim-Toolbox: A MATLAB toolbox for biomechanical motion analysis and creation through simulation'
+tags:
+  - MATLAB
+  - Biomechanics
+  - Movement Simulation
+  - Movement Analysis
+authors:
+  - name: Anne D. Koelewijn  
+    orcid: 0000-0002-1867-0374
+    corresponding: true
+    affiliation: "1, 2"
+  - name: Marlies Nitschke
+    orcid: 0000-0001-7318-9578
+    affiliation: 1
+  - name: Eva Dorschky
+    orcid: 0000-0001-8708-0426
+    affiliation: 1
+  - name: Markus Gambietz
+    orcid: 0000-0003-4096-2354
+    affiliation: 1
+  - name: Alexander Weiss
+    orcid: 0009-0003-1604-6679
+    affiliation: 1
+  - name: Bjoern M. Eskofier
+    orcid: 0000-0002-0417-0336
+    affiliation: 1
+  - name: Antonie (Ton) van den Bogert
+    orcid: 0000-0002-3791-3749
+    affiliation: 2
+affiliations:
+ - name: Friedrich-Alexander-Universität Erlangen-Nürnberg, Erlangen, Germany
+   index: 1
+ - name: Cleveland State University, Cleveland, Ohio, United States of America
+   index: 2
+date: 13 August 2024
+bibliography: paper.bib
+---
+
+# Summary
+We present a MATLAB toolbox for human movement simulation and analysis. The toolbox is versatile due 
+to its object-oriented design, such that different dynamics models can be easily combined with different 
+problem classes. The toolbox's focus is on creating simulations of human walking and running by solving 
+trajectory optimization problems. The trajectory optimization problems are solved using direct collocation. In these problems, the muscle stimulations 
+and initial state are found that minimize an objective related to energy expenditure. Different objectives have been implemented that represent energy or effort minimization. We have also included 
+different objectives for data tracking, including tracking of joint angles, ground reaction forces, accelerometer data, gyroscope data, and marker positions. We have implemented
+two different two dimensional (2D) musucloskeletal models and one three dimensional (3D) musculoskeletal model. 
+To show how the toolbox can be used, we include a 2D and a 3D introductory example, as well as different 
+applications that are based on previous publications. In the future, we aim to implement different other 
+problem classes as well, such as inverse kinematics and dynamics to process experimental data, static and dynamic optimizations to find muscle simulations from joint moments, while also 
+forward shooting can be used to investigate neural control of gait. 
+
+# Statement of Need
+Movements of humans and other animals are extremely versatile, while our efficiency is unmatched by human-made 
+machines without having accurate or efficient controllers. Therefore, a better understanding of human movement can have a large impact for persons with movement disabilities, sports performance optimization, and design of human-made devices. Movement simulations are a key tool for creating this understanding.
+On the one hand, we can use so-called predictive simulations [@ackermann:2010; @falisse:2019;@koelewijn:2018] to investigate unseen movements by replicating the optimization in the central nervous system [@zarrugh:1974] in a computer optimization.
+On the other hand, we can use so-called reconstructive simulations to estimate variables that cannot or were not measured directly by minimizing a data tracking error [@dorscky:2019a;@nitschke:2023;@nitschke:2024]. Such wide-spread measurements and their biomechanical analysis are vital to be able to 
+reach this understanding of human movement.
+
+Here, we present `BioMAC-Sim-Toolbox`, a MATLAB toolbox that can be used to create simulations of human movement 
[…]
