+++ src/problem/@Collocation/writeMovie.m
@@ -1,0 +1,259 @@
+%======================================================================
+%> @file @Collocation/writeMovie.m
+%> @brief Collocation function to create a movie of a simulation result
+%> @details
+%> Details:  Collocation::writeMovie()
+%>
+%> @author Marlies Nitschke
+%> @date January, 2018
+%======================================================================
+
+%======================================================================
+%> @brief Matlab function to create a movie of a simulation result
+%>
+%> @details
+%> Uses Model::showStick() to plot the result.
+%>
+%> Whether the optimal solution is converged is not checked.
+%>
+%> Example usage:
+%> @code
+%> markerTable = result.problem.objectiveTerms(1).varargin{1}.variables; % assuming trackMarker is your first objective
+%> result.problem.writeMovie(result.X, result.filename, [], 0, {1, 0, 1, 0}, 1, 0, markerTable)
+%> @endcode
+%>
+%> @param  obj          Collocation class object
+%> @param  X            Double matrix: State vector (i.e. result) of the problem
+%> @param  filename     (optional) String: Filename which is used to save the OpenSim files.
+%>                      Use an empty char array '' to not save a file.
+%> @param  fps          (optional) Double: Desired frame rate (frames per second)
+%>                      To skip the input, you can use []. (default: 5)
+%> @param  subTrans     (optional) Bool: Subtract the horizontal translation 
+%>                      of the pelvis from the movement. To skip the input, 
+%>                      you can use [].  (default: false)
+%> @param  showStickOpt (optional) Cell: Cell with optional input for model.showStick. Cell can 
+%>                      contain all possible input options after ''range''. (default: empty cell)
+%> @param  makeGIF      (optional) Bool: If true, a gif will be saved.
+%>                      Otherwise an avi will be saved. (default: 0=avi)
+%> @param  playSpeed    (Optional) Double: If specified, the real duration of the simulated motion
+%>                      will be taken into account with
+%>                      - playSpeed == 1: duration of movie reflects the duration of motion
+%>                      - playSpeed  < 1: slow motion (e.g. 0.5 => half the playback speed and double the duration)
+%>                      - playSpeed  > 1: fast motion (e.g. 2 => double the playback speed and half the duration)
+%>                      (default: 0: fps nodes are displayed per second and duration is not taken into account)
+%>                      Attention: The duration of the movie will not be equal to the duration of the motion since
+%>                      the function is plotting the motion till node nNodes and not till nNodes+1.
+%> @param  markerTable  (optional) Table: Variables table specifying markers with at least the columns:
+%>                      type, name, segment, position, direction to call Model.showMarker. If the
+%>                      column "mean" is also give, it will be plotted as reference. (default: empty table)
+%> @param  rangeNodes   (optional) Array: Specifying first and last node
+%>                       which should be used for the movie. This might not work well for symmetric motions
+%>                       which are automatically mirrored.
+%======================================================================
[…]

[…]
