### 2025-02-06 ea352f6
Update main.yml
Typo in docx
