### 2025-02-06 1d3905f
Update build_docs.sh
Change path
