+++ docs/Doxygen/build_docs.sh
@@ -1,6 +1,6 @@
 #!/bin/sh
 
-cp ./Doxyfile ./Doxyfile.back
+cp ./docs/Doxygen/Doxyfile ./docs/Doxygen/Doxyfile.back
 sed -i 's/m=m2cpp.pl/m="perl .\/m2cpp.pl"/' ./Doxyfile
-doxygen ./Doxyfile
-mv ./Doxyfile.back ./Doxyfile
+doxygen ./docs/Doxygen/Doxyfile
+mv ./docs/Doxygen/Doxyfile.back ./docs/Doxygen/Doxyfile
