### 2025-02-06 #7642 @adkoele
@Kevin-Mattheus-Moerman hello again!
The documentation is now working. https://mad-lab-fau.github.io/BioMAC-Sim-Toolbox/

We do not have automatic testing, as I mentioned in the previous comment. I looked into it a bit. If I understand correctly, it would be required to either have automatic testing or to have manual steps that check the correctness of the code. Automatic testing is not really feasible, as the tests in src/tests rely on visual inspection.

How to run tests is explained in the documentation, for example: https://mad-lab-fau.github.io/BioMAC-Sim-Toolbox/class_gait2dc_test.html

Is that sufficient or should I still make an addition to the ReadMe or so?
