### 2025-02-07 #7642 @Kevin-Mattheus-Moerman
@adkoele you are right fully automated testing is not a strict requirement. If you have or could add unit tests (not requiring human interactions) that would be recommended. I think we have enough to proceed here from what you describe but again having automated testing is recommended so perhaps you can consider adding some as this review processes.
