### 2025-03-06 #7642 @prashjha
Howdy! @anmuller, @felixchenier, @romainmartinez, and @tgeijten, would you be interested in reviewing the JOSS submission above ([pdf link](https://github.com/openjournals/joss-papers/blob/joss.07642/joss.07642/10.21105.joss.07642.pdf))? The corresponding software is at this [link](https://github.com/mad-lab-fau/BioMAC-Sim-Toolbox/).

This [link](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) covers almost everything about the review process; I will be happy to help with more information.

Please let me know if you are interested. Thank you!
