### 2025-03-24 #7642 @Aravind-Sundararajan
> Hi [@mprib](https://github.com/mprib) and [@Aravind-Sundararajan](https://github.com/Aravind-Sundararajan), I am looking for another reviewer for the above JOSS submission. Would you be interested in reviewing this? You can find pdf at this link [pdf link](https://github.com/openjournals/joss-papers/blob/joss.07642/joss.07642/10.21105.joss.07642.pdf), and the corresponding software is at this [link](https://github.com/mad-lab-fau/BioMAC-Sim-Toolbox/).
>
> This [reviewer guidelines](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) covers almost everything about the review process; I will be happy to help with more information.
>
> Please let me know if you are interested. Thank you!

Hello @prashjha, I would be happy to review this paper
