### 2025-03-24 #7945 @editorialbot
[REVIEW]: BioMAC-Sim-Toolbox: A MATLAB toolbox for biomechanical motion analysis and creation through simulation
**Submitting author:** @adkoele (<a href="http://orcid.org/0000-0002-1867-0374">Anne Koelewijn</a>)
**Repository:** https://github.com/mad-lab-fau/BioMAC-Sim-Toolbox/
**Branch with paper.md** (empty if default branch):
**Version:** V1.0.1
