### 2025-05-24 mad-lab-fau/BioMAC-Sim-Toolbox#19 @Aravind-Sundararajan
Paper comments
**Organization and Redundancy**

- [ ]     Several concepts (e.g., objective functions, periodicity constraints, solver details) are repeated in different sections with only minor variations. Consider consolidating these into dedicated subsections to improve flow and reduce redundancy.

**Grammar and Clarity**

Numerous grammatical issues and awkward phrasing slightly hinder readability. For instance:

- [ ] “musucloskeletal” -> “musculoskeletal”

- [ ] “weigth” -> “weight”

- [ ] “stimulations” is non-standard, “muscle activations” or “excitation signals” might be more appropriate.

A thorough proofreading pass is recommended to ensure clarity and polish.

**Model Descriptions**

- [ ] The differences between gait2d_osim, gait2dc, and gait3d are scattered throughout the text. A table summarizing key features, origins, dependencies, and publication status of each model would help users choose the most appropriate one for their needs.

**Toolbox Scope and Limitations**

- [ ] While future goals (e.g., inverse dynamics, single shooting) are mentioned, current limitations (e.g., lack of a GUI, steep learning curve for new users, or reliance on IPOPT) are not clearly stated. Adding a short section on “Current Limitations” would increase transparency and set user expectations appropriately.

**User Onboarding**

- [ ] The introductory examples are mentioned but not described in detail. Including a short paragraph or table summarizing what each example does, inputs required, and outputs expected would help onboard new users more effectively.

**Performance and Benchmarks**

- [ ] While compute times are provided for 2D and 3D models, it would be helpful to specify the hardware used for benchmarking and possibly include memory usage or solver settings to contextualize these performance claims.

**Licensing and Availability**

- [ ] The manuscript should clearly state the license type (e.g., MIT, GPL) under which the toolbox is released, and where the code can be accessed (e.g., GitHub link, Zenodo DOI). This is especially important for reproducibility and community adoption.

This issue is part of a [JOSS review](https://github.com/openjournals/joss-reviews/issues/7945).
