### 2025-08-14 d9018d2
Update README.md
Updated according to the suggestions in #20
