+++ paper.bib
@@ -39,7 +39,39 @@
   pages={1055--1060},
   year={2010},
   publisher={Elsevier},
+  publisher={Elsevier},
   doi = {10.1016/j.jbiomech.2009.12.012}
+}
+
+@inproceedings{dhondt:2024,
+  title={PredSim: a framework for rapid predictive simulations of locomotion},
+  author={D'hondt, Lars and Falisse, Antoine and Gupta, Dhruv and Van Den Bosch, Bram and Buurke, Tom JW and Febrer-Nafr{\'\i}a, M{\'\i}riam and Vandekerckhove, Ines and Afschrift, Maarten and De Groote, Friedl},
+  booktitle={2024 10th IEEE RAS/EMBS International Conference for Biomedical Robotics and Biomechatronics (BioRob)},
+  pages={1208--1213},
+  year={2024},
+  organization={IEEE}
+}
+
+@article{michaud:2022,
+  title={Bioptim, a python framework for musculoskeletal optimal control in biomechanics},
+  author={Michaud, Benjamin and Bailly, Fran{\c{c}}ois and Charbonneau, Eve and Ceglia, Amedeo and Sanchez, L{\'e}a and Begon, Mickael},
+  journal={IEEE Transactions on Systems, Man, and Cybernetics: Systems},
+  volume={53},
+  number={1},
+  pages={321--332},
+  year={2022},
+  publisher={IEEE}
+}
+
+@article{dembia:2020,
+  title={Opensim moco: Musculoskeletal optimal control},
+  author={Dembia, Christopher L and Bianco, Nicholas A and Falisse, Antoine and Hicks, Jennifer L and Delp, Scott L},
+  journal={PLOS Computational Biology},
+  volume={16},
+  number={12},
+  pages={e1008493},
+  year={2020},
+  publisher={Public Library of Science San Francisco, CA USA}
 }
 
 @article{nitschke:2023,
@@ -144,14 +176,14 @@
   doi = {10.1016/j.gaitpost.2016.07.007}
 }
 
-@article{dorschkynitschke:2024,
+@article{dorschkynitschke:2025,
   title={Comparing sparse inertial sensor setups for sagittal-plane walking and running reconstructions},
   author={Dorschky, Eva and Nitschke, Marlies and Mayer, Matthias and Weygers, Ive and Gassner, Heiko and Seel, Thomas and Eskofier, Bjoern M and Koelewijn, Anne D},
-  journal={bioRxiv},
-  pages={2023--05},
-  year={2024},
-  publisher={Cold Spring Harbor Laboratory},
-  doi = {10.1101/2023.05.25.542228}
+  journal={Frontiers in Bioengineering and Biotechnology},
+  volume={13},
+  pages={1507162},
+  year={2025},
+  publisher={Frontiers Media SA}
 }
 
 @article{dorschky:2019b,
