### 2025-08-15 f85f0d4
Updated references, completed table
