+++ paper.bib
@@ -19,15 +19,24 @@
   doi = {10.1007/BF00430237}
 }
 
-@article{seth:2011,
-  title={OpenSim: a musculoskeletal modeling and simulation framework for in silico investigations and exchange},
-  author={Seth, Ajay and Sherman, Michael and Reinbolt, Jeffrey A and Delp, Scott L},
-  journal={Procedia Iutam},
-  volume={2},
-  pages={212--232},
-  year={2011},
-  publisher={Elsevier},
-  doi = {10.1016/j.piutam.2011.04.021}
+@inproceedings{gambietz:2024,
+  title={Unobtrusive Gait Reconstructions Using Radar-Based Optimal Control Simulations},
+  author={Gambietz, Markus and Dr{\"o}ge, Anne and Sch{\"u}{\ss}ler, Christian and Stahlke, Maximilian and Wirth, Vanessa and Miehling, J{\"o}rg and Vossiek, Martin and Koelewijn, Anne D},
+  booktitle={2024 58th Asilomar Conference on Signals, Systems, and Computers},
+  pages={1505--1509},
+  year={2024},
+  organization={IEEE}
+}
+
+@article{seth:2018,
+  title={OpenSim: Simulating musculoskeletal dynamics and neuromuscular control to study human and animal movement},
+  author={Seth, Ajay and Hicks, Jennifer L and Uchida, Thomas K and Habib, Ayman and Dembia, Christopher L and Dunne, James J and Ong, Carmichael F and DeMers, Matthew S and Rajagopal, Apoorva and Millard, Matthew and others},
+  journal={PLoS computational biology},
+  volume={14},
+  number={7},
+  pages={e1006223},
+  year={2018},
+  publisher={Public Library of Science San Francisco, CA USA}
 }
 
 @article{ackermann:2010,
