### 2025-08-15 4c61a3a
ExoWalking Update
Changed name of script to be consistent
