+++ Tutorial/+ISB_Tutorial/move2D_IMU.m
@@ -1,0 +1,101 @@
+%======================================================================
+%> @file +ISB_tutorial/move2D_IMU.m
+%> @brief Function to set up optimal control problem to simulate walking or
+%> running from inertial sensor data
+%>
+%> @author Eva Dorschky, Anne Koelewijn
+%> @date November, 2024
+%======================================================================
+
+% ======================================================================
+%> @brief Function to specify the optimization problem for 2D running, used
+%> in the 2025 ISB tutorial
+%>
+%> @param   model          Gait2dc: Model which should be used for the simulation
+%> @param   resultfile     String: Name of the resultfile including path
+%> @param   trackingData   TrackingData: Tracking Data containing angles and GRFs data 
+%> @param   initialGuess   String: Filename with path specifying the initial guess 
+%> @param   isSymmetric    Bool: Specifies if we assume movement symmetry. If so, 
+%>                         we simulate only one half gait cycle. This has to fit to the tracking data.
+%> @param   W              Struct: objective weigths
+%> @param   targetSpeed    Double: Target speed of the movement in x direction in m/s. 
+%>                         This speed will be enforced.
+%> @param   targetdur      Double: Target durection of the movement in s. 
+%>                         This duration will be enforced.
+%> @retval  problem        Collocation: Optimization problem for 2D running
+% ======================================================================
+function problem = move2D_IMU(model,resultfile, trackingData, initialGuess, varargin)
+
+if isempty(varargin)
+    W.track  = 1;                 % Weight of tracking term in objective
+    W.eff    = 600;               % Weight of effort term  in objective
+    W.reg    = 1e-3;              % Weight of regularization term in objective
+    N = 100;
+else
+    for i = 1:length(varargin)
+        if isa(varargin{i}, 'struct')
+            W = varargin{i};
+        else
+            W.track  = 1;                 % Weight of tracking term in objective
+            W.eff    = 600;               % Weight of effort term  in objective
+            W.reg    = 1e-3;              % Weight of regularization term in objective
+        end
+        
+        if isa(varargin{i}, 'double')
+            N = varargin{i};
+        else
+            N = 100;
+        end
+    end
+end
+
+Euler = 'BE'; % we use backward Euler for discretization
+isSymmetric = 0; % we assume no symmetry
+plotLog = 1; % Show progress of optimization
+
+problem = Collocation(model,N,Euler,resultfile,plotLog);
+
+% Tracking data preprocessing
+if trackingData.nSamples~=N
+    trackingData.resampleData(N);
+end
+targetspeed =  trackingData.variables.mean{strcmp(trackingData.variables.type,'speed')};
+
+% Create optimization variables
+xmin = model.states.xmin;
+xmax = model.states.xmax;
+
+states_min = repmat(xmin,1,N+1); %We add N+1 states because we apply the periodicity constraint on the (N+1)th node, which should be the same as the first
+states_max = repmat(xmax,1,N+1);
+
+idxPelvisX = model.extractState('q', 'pelvis_tx');
+states_min(idxPelvisX,1) = 0; % trunk position in x direction is zero in first node to avoid an infinite amount of possible solutions
+states_max(idxPelvisX,1) = 0; % trunk position in x direction is zero in first node to avoid an infinite amount of possible solutions
+
+problem.addOptimVar('states',states_min,states_max);
+problem.addOptimVar('controls',repmat(model.controls.xmin,1,N+1),repmat(model.controls.xmax,1,N+1));
+
[…]
