### 2025-08-15 8adcafd
Update speedConstraint.m
Removed commented lines that weren't used
