### 2025-08-21 ad1ed12
Updated paper and documentation based on comments
