+++ paper.bib
@@ -176,7 +176,7 @@
 
 @article{koelewijn:2016,
   title={Joint contact forces can be reduced by improving joint moment symmetry in below-knee amputee gait simulations},
-  author={Koelewijn, Anne D and van den Bogert, Antonie J},
+  author={Koelewijn, Anne D and Van den Bogert, Antonie J},
   journal={Gait \& Posture},
   volume={49},
   pages={219--225},
