+++ paper.md
@@ -25,7 +25,7 @@ authors:
   - name: Bjoern M. Eskofier
     orcid: 0000-0002-0417-0336
     affiliation: 1
-  - name: Antonie (Ton) van den Bogert
+  - name: Antonie J. van den Bogert
     orcid: 0000-0002-3791-3749
     affiliation: 2
 affiliations:
@@ -38,16 +38,16 @@ bibliography: paper.bib
 ---
 
 # Summary
-We present a MATLAB toolbox for human movement simulation and analysis. The toolbox is versatile due 
-to its object-oriented design, such that different dynamics models can be easily combined with different 
-problem classes. The toolbox's focus is on creating simulations of human walking and running by solving 
+We present a MATLAB toolbox for human movement simulation and analysis. The toolbox's focus is on creating simulations of human walking and running by solving 
 trajectory optimization problems. **In these problems, the muscle excitations  and initial state are found that minimize an objective related to energy expenditure. **
 To solve the trajectory optimization problems, we have implemented a direct collocation framework.
 Different objectives have been implemented that represent energy or effort minimization. We have also included 
 different objectives for data tracking, including tracking of joint angles, ground reaction forces, accelerometer data, gyroscope data, and marker positions. We have implemented
 two different two dimensional (2D) **musculoskeletal** models and one three dimensional (3D) musculoskeletal model. 
-To show how the toolbox can be used, we include a 2D and a 3D introductory example, as well as different 
-applications that are based on previous publications. In the future, we aim to implement different other 
+To show how the toolbox can be used, we include a **tutorial**, a 2D and a 3D introductory example, as well as different 
+applications that are based on previous publications. The toolbox is versatile due 
+to its object-oriented design, such that different dynamics models can be easily combined with different 
+problem classes. In the future, we aim to implement different other 
 problem classes as well, such as inverse kinematics and dynamics to process experimental data, static and dynamic optimizations to find muscle simulations from joint moments, while also 
 forward shooting can be used to investigate neural control of gait. 
 
@@ -57,38 +57,38 @@ machines without having accurate or efficient controllers. Therefore, a better u
 and design of human-made devices. Movement simulations are a key tool for creating this understanding.
 On the one hand, we can use so-called predictive simulations [@ackermann:2010; @falisse:2019;@koelewijn:2018] to investigate unseen movements by replicating the optimization in the central nervous system [@zarrugh:1974] in a computer optimization. **Such simulations can help us decipher fundamentals of movement **
 **in humans and other animals.**
-On the other hand, we can use so-called reconstructive simulations to estimate variables that cannot or were not measured directly by minimizing a data tracking error [@dorscky:2019a;@nitschke:2023;@nitschke:2024]. **These reconstructive simulations do not require laboratory equipment, **
+On the other hand, we can use so-called reconstructive simulations to estimate variables that cannot or were not measured directly by minimizing a data tracking error [@dorscky:2019a;@nitschke:2023;@nitschke:2024]. **These reconstructive simulations do not necessarily require laboratory equipment, **
 **instead allowing biomechanical variables to be estimated accurately from wearable or contactless sensors, such as inertial sensors. Wide-spread measurement of movements and their biomechanical analysis are vital to be able to **
 **reach a deep understanding** of human movement.
 
-Here, we present `BioMAC-Sim-Toolbox`, a MATLAB toolbox that can be used to create simulations of human movement 
[…]
