### 2025-08-21 e22a276
Added headers to paper
