### 2025-08-21 bf79710
Update paper.md
Small bug fix
