### 2025-08-21 #7945 @adkoele
Hi @rmndrs89!

I am working on the resubmission right now and will summarise the changes that we have made here. It might take me a bit more time to finish the commits though. First of all, thank you for the review, your comments were helpful. I apologize for the delay in addressing them.

- Functionality: we have also identified the Linux issue. Unfortunately, it is not something that we can fix ourselves, so for now, Linux users have to use gcc version 9 or older.
- Documentation: I recently gave a tutorial at the conference of the International Society of Biomechanics and I have added it to the toolbox to get started. This tutorial is partly a tutorial and partly a how-to guide according to the Diataxis framework, because it tells the user what to do but also ensures that the task can be completed. Furthermore, it provides an entry-level example as well.
- Community guidelines on how to contribute can be found here: https://github.com/mad-lab-fau/BioMAC-Sim-Toolbox/blob/main/CONTRIBUTING.md
- Statement of need: I have added a link to the JOSS paper on the main page of the documentation. Note that it is not yet working, as the paper is not published.
- We have added a paragraph comparing to existing software suits in lines 162-173

I believe that we have now addressed all your comments. If anything else is missing, please let us know!
