### 2025-08-21 f7821db
Added IPOPT limitation
