+++ paper.bib
@@ -39,6 +39,18 @@
   publisher={Public Library of Science San Francisco, CA USA}
 }
 
+@incollection{bueskens:2013,
+  author = {Büskens, Christof and Wassel, Dennis},
+  editor = {Fasano, Giorgio and Pintér, János D.},
+  title = {The ESA NLP Solver WORHP},
+  booktitle = {Modeling and Optimization in Space Engineering},
+  publisher = {Springer New York},
+  year = {2013},
+  volume = {73},
+  pages = {85--110},
+  doi = {10.1007/978-1-4614-4469-5_4}
+}
+
 @article{ackermann:2010,
   title={Optimality principles for model-based prediction of human gait},
   author={Ackermann, Marko and Van den Bogert, Antonie J},
