### 2025-08-21 mad-lab-fau/BioMAC-Sim-Toolbox#20 @adkoele
Thanks for these suggestions, they were very helpful. We have addressed them. Would you want us to tick the boxes or do you want to check yourself?
