### 2025-10-15 9a113e5
Adding flexibility to Gait3d models
Changes to allow models to be stored in other locations than the model file.
