### 2025-10-15 9009c43
Update IPOPT.m
Only checks for Jacobian patterns when there are constraints.
