### 2025-10-15 795944e
Muscle dynamics
Aligning with previous commit related to models in other folders
