### 2025-10-15 e973d77
Increased memory allocation for dM_dx
354 was required instead of 352. Also changed for the dependent variables.
