### 2025-10-16 2c2b5f2
Debugging for 3d models
Added functionality to directly input external torques (as they're used in the arms). It is working without error but not properly checked yet.
