### 2025-10-27 #7945 @prashjha
Hi @adkoele, some more action items for you.

- Ensure the Zenodo archive title and author list are the same as in the Joss draft
- Confirm that the version is v1.0.1
- Check and/or fix the missing dois warnings/errors above. Let me know if the references do not have DOIs so I can ignore these.
